* 1.5.2   - コンパイラをVisual Studio 2013 (VC12)に変更   - ワーニング除去   - Windows8でプラグインが固まる問題の対応   - afxdrive 0.1.2     ドライブ名に")"が含まれると正しく開けない問題の修正

+++ afxmenu/afxmenu.cpp
@@ -172,7 +172,7 @@ void _line2menu(wchar_t* line, MenuItem& item)
 		} else {
 			if (*p != L' ' && *p != L'\t') {
 				wcstombs(item.cmd, p, sizeof(item.cmd));
-				int len = wcslen(p);
+				size_t len = wcslen(p);
 				if (p[len-2] == L'\n') {
 					p[len-2] = L'\0';
 				}
